refactor: move database code to a new db.py file in zerokdbapi directory
